fix(doctor): scrub all PII from output, atomic WAV write, gate fixes
Gate and review findings on the orchestration/listen/report layer.

PII: the report and walkthrough are posted publicly, so no credentials,
host, or IP may appear. redactTarget strips userinfo, host, port and
query from the target line (previously rtsp.RedactURL, which kept the
host); a piiScrubber removes the host and any IP literal from a failed
step's detail in the walkthrough, and a malformed-URL error (which the
standard library echoes with credentials) is reduced to a category.

WAV listen: write to a temp file and rename into place only on success,
so a decode or write failure, an interrupted run, or an unsupported codec
never truncates, clobbers, or leaves a partial file at the --wav path; the
Close error is checked rather than deferred and discarded.

Also: the walkthrough surfaces the listen result on the default path, not
only under --report; a PLAY failure reports "play failed" rather than the
copied "connection failed"; finish's result phrase ordering matches
mapExitClean; the AAC frame count is computed in int64 so a >2 GiB decode
does not wrap on 32-bit builds; two comments drop internal planning vocab.

Tests: redact unit and end-to-end, sanitizeWriteErr, the pre-capture
failure table, the listen skip and write paths, and EndReason.String.

+++ cmd/stream-doctor/internal/doctor/listen.go
@@ -179,10 +179,12 @@ func writeWAVAAC(w io.Writer, track rtsp.Track, frames []CapturedFrame) (ListenR
 		return ListenResult{}, err
 	}
 
-	bytesPerFrame := 2 * info.Channels
+	bytesPerFrame := int64(2 * info.Channels)
 	var frameCount int
 	if bytesPerFrame > 0 {
-		frameCount = int(written) / bytesPerFrame
+		// written is int64; divide before narrowing so a >2 GiB decode does
+		// not wrap on 32-bit builds. The quotient (a frame count) fits int.
+		frameCount = int(written / bytesPerFrame)
 	}
 
 	return ListenResult{
